requirements.txt
matplotlib==3.4.2
pandas==1.2.4
numpy==1.20.2
Pillow==8.3.1
scikit_learn==0.24.2
statsmodels==0.12.2
streamlit==0.86.0
xgboost==1.4.2
